Safari-609.1.11.1/LayoutTests/crypto/subtle/aes-cfb-generate-export-key-jwk-length-256-expected.txt

Test exporting a 256 bits AES-CFB key with JWK format

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Generating a key...
Exporting a key...
PASS key.kty is 'oct'
PASS key.key_ops is ['decrypt', 'encrypt', 'unwrapKey', 'wrapKey']
PASS key.alg is 'A256CFB8'
PASS key.ext is true
PASS Base64URL.parse(key.k).byteLength is 32
PASS successfullyParsed is true

TEST COMPLETE
